requirements.txt
blinker==1.7.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Flask==3.0.2
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.57.0
greenlet==3.0.3
imbalanced-learn==0.13.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sklearn-compat==0.1.3
SQLAlchemy==2.0.28
threadpoolctl==3.6.0
typing_extensions==4.10.0
tzdata==2025.2
Werkzeug==3.0.1
WTForms==3.2.1
seaborn
reportlab
